requirements.txt
apache-airflow-providers-docker==3.14.0
apache-airflow-providers-google==10.23.0
apache-airflow-providers-slack==8.9.0
apache-airflow==2.10.2
google-cloud-bigquery==3.26.0
google-cloud-storage==2.18.2
requests==2.32.3
